requirements-dev.txt
pyyaml>=6.0,<7.0
jsonschema>=4.18,<5.0
Pillow>=10.0,<11.0
